+++ src/main/java/com/kronoly/DTO/ClienteResumoDTO.java
@@ -9,9 +9,9 @@ public class ClienteResumoDTO {
     private int idCliente;
     private String documento;
     public String nome;
-    public int telefone;
+    public Long telefone;
 
-    public ClienteResumoDTO(int idCliente, String documento, String nome, int telefone) {
+    public ClienteResumoDTO(int idCliente, String documento, String nome, Long telefone) {
         this.idCliente = idCliente;
         this.documento = documento;
         this.nome = nome;
