+++ src/main/java/com/kronoly/DTO/UsuarioUpdateDTO.java
@@ -12,7 +12,7 @@ public class UsuarioUpdateDTO {
     private String nome;
     private String email;
     private String senha;
-    private int telefone;
+    private Long telefone;
     private TipoUsuarioEnum tipoUsuario;
     private String caminhoFoto;
